+++ HISTORY.md
@@ -55,6 +55,7 @@
 
 ### 2019-04-02
 * Objects, Threads: lazily initialize volume property
+* Objects: use AudioContext to play recorded sounds
 
 ### 2019-04-01
 * Objects: let the Microphone share the Note prototype's AudioContext